pyproject.toml
# see documentation, e.g.
# - https://packaging.python.org/en/latest/specifications/declaring-project-metadata/#declaring-project-metadata
# - https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
# - https://www.python.org/dev/peps/pep-0621/

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["."] by default)
exclude = []  # exclude packages matching these glob patterns (empty by default)
namespaces = false  # to disable scanning PEP 420 namespaces (true by default)

[project]
authors = [

    { name = "Anas Alhariri", email = "anas.alhariri@study.hs-duesseldorf.de" },
    { name = "Jonah Gräfe", email = "jonah.graefe@study.hs-duesseldorf.de" }
]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "ruff == 0.1.7",
    "pytest", 
    "numpy==1.26.1",
    "pandas==2.2.0",
    "torch",
    "torchvision",
    "ultralytics==8.2.46",
    "tqdm",
    "Pillow",
    "opencv-python",
    "scikit-learn",
    "scipy",
    "flask",
    "waitress",
]

description = "A package implementing the Image-recommender"
license = {file = "LICENSE"}
name = "image-recommender"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.8"
version = "0.1.0"

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "isort",
    "pytest",
    "pytest-cov",
    "ruff",
]
publishing = [
    "build",
    "twine",
    "wheel",
]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
)/
'''

[tool.isort]
profile = "black"
known_first_party = "image_recommender"
packages = ["python_image_recommender"]
src_paths = ["image-recommender", "tests"]
line_length = "100"

[tool.pylint]  # If you opt to use pylint
# Pylint configuration options
